Updated project

+++ javaparalleldemo/src/main/java/com/github/jeandersonbc/javaparalleldemo/arraysum/ArraySumProblem.java
@@ -2,6 +2,6 @@ package com.github.jeandersonbc.javaparalleldemo.arraysum;
 
 public interface ArraySumProblem {
 
-	int sum(Integer[] input);
+	Double sum(Double[] input);
 
 }
